Time: 8 ms (96.33%), Space: 49.1 MB (74.16%) - LeetHub

+++ 1161-maximum-level-sum-of-a-binary-tree/1161-maximum-level-sum-of-a-binary-tree.java
@@ -22,15 +22,15 @@ class Solution {
         int ansLevel = 0;
         while(q.size() > 0){
             int n = q.size();
-            long sum = 0;
+            int sum = 0;
             for(int i = 0; i<n; i++){
                 TreeNode temp = q.remove();
                 sum += temp.val;
                 if(temp.left != null) q.add(temp.left);
                 if(temp.right != null) q.add(temp.right);
             }
-            if((int)sum>maxSum){
-                maxSum = (int)sum;
+            if(sum>maxSum){
+                maxSum = sum;
                 ansLevel = level;
             }
             level++;
